requirements.txt
paddlepaddle
paddleocr
PyMuPDF
Pillow
numpy
pytesseract